requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
Pillow==10.1.0
numpy==1.26.2
python-jose==3.3.0
pytest==7.4.3
black==23.11.0
